[Fabric-Bridge] Refactor device management from main
